fix(functions): fix post structural QC dir check
In current version, when running -proc-func with MCS dataset,
the error is thrown, saying that proc_posprocessing was not run.

This might be because in MSC structural scans were done separately
to other scans (separate sessions), so when running functional sessions
QC files are checked in `dir_QC=$out/${subject}/${SES}/QC `, in spite of
providing `-antStruct` directory.

Proposed solution is to move the check on QC files after the section,
where variables using `-antStruct` are generated and add 2 variables that would
point to the QC directory in the structural directory.